+++ application/Espo/Classes/FieldValidators/ArrayType.php
@@ -27,25 +27,40 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
 use Espo\ORM\Entity;
 
-class PersonNameType extends BaseType
+class ArrayType extends BaseType
 {
     public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
     {
-        $isEmpty = true;
-        foreach ($this->getActualAttributeList($entity, $field) as $attribute) {
-            if ($attribute === 'salutation' . ucfirst($field)) {
-                continue;
-            }
-            if ($entity->has($attribute) && $entity->get($attribute) !== '') {
-                $isEmpty = false;
-                break;
-            }
-        }
-        if ($isEmpty) return false;
+        return $this->isNotEmpty($entity, $field);
+    }
+
+    public function checkMaxCount(Entity $entity, string $field, $validationValue, $data) : bool
+    {
+        if (!$this->isNotEmpty($entity, $field)) return true;
+        $list = $entity->get($field);
+        if (count($list) > $validationValue) return false;
         return true;
+    }
+
+    public function checkArray(Entity $entity, string $field, $validationValue, $data) : bool
+    {
+        if (isset($data->$field) && $data->$field !== null && !is_array($data->$field)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    protected function isNotEmpty(Entity $entity, $field)
+    {
+        if (!$entity->has($field) || $entity->get($field) === null) return false;
+        $list = $entity->get($field);
+        if (!is_array($list)) return false;
+        if (count($list)) return true;
+        return false;
     }
 }
