+++ application/Espo/Classes/FieldValidators/BaseType.php
@@ -27,14 +27,38 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
 use Espo\ORM\Entity;
 
-class LinkMultipleType extends BaseType
+use Espo\Core\{
+    Utils\Metadata,
+    Utils\FieldManagerUtil,
+};
+
+class BaseType
 {
-    public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
+    protected $metadata;
+    protected $fieldManagerUtil;
+
+    public function __construct(Metadata $metadata, FieldManagerUtil $fieldManagerUtil)
     {
-        return count($entity->getLinkMultipleIdList($field)) > 0;
+        $this->metadata = $metadata;
+        $this->fieldManagerUtil = $fieldManagerUtil;
+    }
+
+    protected function getActualAttributeList(Entity $entity, string $field) : array
+    {
+        return $this->getFieldManagerUtil()->getActualAttributeList($entity->getEntityType(), $field);
+    }
+
+    protected function getMetadata() : Metadata
+    {
+        return $this->metadata;
+    }
+
+    protected function getFieldManagerUtil() : FieldManagerUtil
+    {
+        return $this->fieldManagerUtil;
     }
 }
