+++ application/Espo/Classes/FieldValidators/DateType.php
@@ -27,51 +27,19 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
 use Espo\ORM\Entity;
 
-class EmailType extends BaseType
+class DateType extends BaseType
 {
     public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
     {
-        if ($this->isNotEmpty($entity, $field)) return true;
-
-        $dataList = $entity->get($field . 'Data');
-        if (!is_array($dataList)) return false;
-
-        foreach ($dataList as $item) {
-            if (!empty($item->emailAddress)) return true;
-        }
-
-        return false;
-    }
-
-    public function checkEmailAddress(Entity $entity, string $field, $validationValue, $data) : bool
-    {
-        if ($this->isNotEmpty($entity, $field)) {
-            $address = $entity->get($field);
-            if (!filter_var($address, FILTER_VALIDATE_EMAIL)) {
-                return false;
-            }
-        }
-
-        $dataList = $entity->get($field . 'Data');
-        if (is_array($dataList)) {
-            foreach ($dataList as $item) {
-                if (empty($item->emailAddress)) continue;
-                $address = $item->emailAddress;
-                if (!filter_var($address, FILTER_VALIDATE_EMAIL)) {
-                    return false;
-                }
-            }
-        }
-
-        return true;
+        return $this->isNotEmpty($entity, $field);
     }
 
     protected function isNotEmpty(Entity $entity, $field)
     {
-        return $entity->has($field) && $entity->get($field) !== '' && $entity->get($field) !== null;
+        return $entity->has($field) && $entity->get($field) !== null;
     }
 }
