+++ application/Espo/Classes/FieldValidators/FileType.php
@@ -27,8 +27,8 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
-class FloatType extends IntType
+class FileType extends LinkType
 {
 }
