+++ application/Espo/Classes/FieldValidators/ImageType.php
@@ -27,30 +27,8 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
-use Espo\ORM\Entity;
-
-class VarcharType extends BaseType
+class ImageType extends FileType
 {
-    public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
-    {
-        return $this->isNotEmpty($entity, $field);
-    }
-
-    public function checkMaxLength(Entity $entity, string $field, $validationValue, $data) : bool
-    {
-        if ($this->isNotEmpty($entity, $field)) {
-            $value = $entity->get($field);
-            if (mb_strlen($value) > $validationValue) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    protected function isNotEmpty(Entity $entity, $field)
-    {
-        return $entity->has($field) && $entity->get($field) !== '' && $entity->get($field) !== null;
-    }
 }
