+++ application/Espo/Classes/FieldValidators/IntType.php
@@ -27,8 +27,33 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
-class ImageType extends FileType
+use Espo\ORM\Entity;
+
+class IntType extends BaseType
 {
+    public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
+    {
+        return $this->isNotEmpty($entity, $field);
+    }
+
+    public function checkMax(Entity $entity, string $field, $validationValue, $data) : bool
+    {
+        if (!$this->isNotEmpty($entity, $field)) return true;
+        if ($entity->get($field) > $validationValue) return false;
+        return true;
+    }
+
+    public function checkMin(Entity $entity, string $field, $validationValue, $data) : bool
+    {
+        if (!$this->isNotEmpty($entity, $field)) return true;
+        if ($entity->get($field) < $validationValue) return false;
+        return true;
+    }
+
+    protected function isNotEmpty(Entity $entity, $field)
+    {
+        return $entity->has($field) && $entity->get($field) !== null;
+    }
 }
