+++ application/Espo/Classes/FieldValidators/JsonArrayType.php
@@ -27,28 +27,25 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
 use Espo\ORM\Entity;
 
-class PhoneType extends BaseType
+class JsonArrayType extends BaseType
 {
-    public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
+    public function checkArray(Entity $entity, string $field, $validationValue, $data) : bool
     {
-        if ($this->isNotEmpty($entity, $field)) return true;
+        if (!$entity->has($field) || $entity->get($field) === null) return true;
 
-        $dataList = $entity->get($field . 'Data');
-        if (!is_array($dataList)) return false;
-
-        foreach ($dataList as $item) {
-            if (!empty($item->phoneNumber)) return true;
-        }
-
-        return false;
+        return is_array($entity->get($field));
     }
 
     protected function isNotEmpty(Entity $entity, $field)
     {
-        return $entity->has($field) && $entity->get($field) !== '' && $entity->get($field) !== null;
+        if (!$entity->has($field) || $entity->get($field) === null) return false;
+        $list = $entity->get($field);
+        if (!is_array($list)) return false;
+        if (count($list)) return true;
+        return false;
     }
 }
