+++ application/Espo/Classes/FieldValidators/PersonNameType.php
@@ -27,9 +27,25 @@
  * these Appropriate Legal Notices must retain the display of the "EspoCRM" word.
  ************************************************************************/
 
-namespace Espo\Core\FieldValidators;
+namespace Espo\Classes\FieldValidators;
 
-class ArrayIntType extends ArrayType
+use Espo\ORM\Entity;
+
+class PersonNameType extends BaseType
 {
-
+    public function checkRequired(Entity $entity, string $field, $validationValue, $data) : bool
+    {
+        $isEmpty = true;
+        foreach ($this->getActualAttributeList($entity, $field) as $attribute) {
+            if ($attribute === 'salutation' . ucfirst($field)) {
+                continue;
+            }
+            if ($entity->has($attribute) && $entity->get($attribute) !== '') {
+                $isEmpty = false;
+                break;
+            }
+        }
+        if ($isEmpty) return false;
+        return true;
+    }
 }
