+++ application/Espo/Resources/metadata/app/templateHelpers.json
@@ -1,3 +1,3 @@
 {
-    "googleMapsImage": "\\Espo\\Core\\TemplateHelpers\\GoogleMaps::image"
+    "googleMapsImage": "Espo\\Classes\\TemplateHelpers\\GoogleMaps::image"
 }
